#279 cleanup transcoding class hierarchy
